+++ modules/welcomer/models/User.js
@@ -1,0 +1,26 @@
+const {DataTypes, Model} = require('sequelize');
+
+module.exports = class User extends Model {
+    static init(sequelize) {
+        return super.init({
+            id: {
+                autoIncrement: true,
+                type: DataTypes.INTEGER,
+                primaryKey: true
+            },
+            userID: DataTypes.STRING,
+            channelID: DataTypes.STRING,
+            messageID: DataTypes.STRING,
+            timestamp: DataTypes.DATE
+        }, {
+            tableName: 'welcomer_User',
+            timestamps: true,
+            sequelize
+        });
+    }
+};
+
+module.exports.config = {
+    'name': 'User',
+    'module': 'welcomer'
+};